+++ packages/react-components/react-nav-preview/stories/Hamburger/HamburgerDefault.stories.tsx
@@ -1,0 +1,4 @@
+import * as React from 'react';
+import { Hamburger, HamburgerProps } from '@fluentui/react-nav-preview';
+
+export const Default = (props: Partial<HamburgerProps>) => <Hamburger {...props} />;