+++ packages/react-components/react-tag-picker-preview/stories/TagPicker/TagPickerDefault.stories.tsx
@@ -8,7 +8,7 @@ import {
   TagPickerOption,
   TagPickerGroup,
 } from '@fluentui/react-tag-picker-preview';
-import { Tag, Avatar } from '@fluentui/react-components';
+import { Tag, Avatar, Field } from '@fluentui/react-components';
 
 const options = [
   'John Doe',
@@ -26,9 +26,10 @@ export const Default = () => {
   const onOptionSelect: TagPickerProps['onOptionSelect'] = (e, data) => {
     setSelectedOptions(data.selectedOptions);
   };
+  const tagPickerOptions = options.filter(option => !selectedOptions.includes(option));
 
   return (
-    <div style={{ maxWidth: 400 }}>
+    <Field label="Select Employees" style={{ maxWidth: 400 }}>
       <TagPicker onOptionSelect={onOptionSelect} selectedOptions={selectedOptions}>
         <TagPickerControl>
           <TagPickerGroup>
@@ -46,20 +47,20 @@ export const Default = () => {
           <TagPickerInput aria-label="Select Employees" />
         </TagPickerControl>
         <TagPickerList>
-          {options
-            .filter(option => !selectedOptions.includes(option))
-            .map(option => (
-              <TagPickerOption
-                secondaryContent="Microsoft FTE"
-                media={<Avatar aria-hidden name={option} color="colorful" />}
-                value={option}
-                key={option}
-              >
-                {option}
-              </TagPickerOption>
-            ))}
+          {tagPickerOptions.length > 0
+            ? tagPickerOptions.map(option => (
+                <TagPickerOption
+                  secondaryContent="Microsoft FTE"
+                  media={<Avatar shape="square" aria-hidden name={option} color="colorful" />}
+                  value={option}
+                  key={option}
+                >
+                  {option}
+                </TagPickerOption>
+              ))
+            : 'No options available'}
         </TagPickerList>
       </TagPicker>
-    </div>
+    </Field>
   );
 };