+++ packages/react-components/react-shared-contexts/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-shared-contexts
 
-This log was last generated on Tue, 23 Apr 2024 08:17:48 GMT and should not be manually modified.
+This log was last generated on Mon, 06 May 2024 12:55:00 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.18.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-shared-contexts_v9.18.0)
+
+Mon, 06 May 2024 12:55:00 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-shared-contexts_v9.17.0..@fluentui/react-shared-contexts_v9.18.0)
+
+### Minor changes
+
+- feat: adds TagPicker to custom style hooks ([PR #31272](https://github.com/microsoft/fluentui/pull/31272) by bernardo.sunderhus@gmail.com)
 
 ## [9.17.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-shared-contexts_v9.17.0)