+++ packages/react-components/react-switch/CHANGELOG.md
@@ -1,8 +1,23 @@
 # Change Log - @fluentui/react-switch
 
-This log was last generated on Thu, 02 May 2024 11:36:38 GMT and should not be manually modified.
+This log was last generated on Mon, 06 May 2024 12:55:01 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.1.80](https://github.com/microsoft/fluentui/tree/@fluentui/react-switch_v9.1.80)
+
+Mon, 06 May 2024 12:55:01 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-switch_v9.1.79..@fluentui/react-switch_v9.1.80)
+
+### Patches
+
+- fix: Add HC styling for Switch font icon ([PR #31241](https://github.com/microsoft/fluentui/pull/31241) by ololubek@microsoft.com)
+- Bump @fluentui/react-field to v9.1.64 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-jsx-runtime to v9.0.37 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-label to v9.1.69 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-shared-contexts to v9.18.0 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-tabster to v9.21.1 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
+- Bump @fluentui/react-utilities to v9.18.8 ([PR #31271](https://github.com/microsoft/fluentui/pull/31271) by beachball)
 
 ## [9.1.79](https://github.com/microsoft/fluentui/tree/@fluentui/react-switch_v9.1.79)